requirements.txt
pytest==9.0.3
selenium==4.23.1
Appium-Python-Client==4.3.0